LayoutTests/imported/w3c/web-platform-tests/css/css-masking/hit-test/clip-path-element-objectboundingbox-002-expected.txt


FAIL Hit-test of clip-path nested objectBoundingBox <clipPath> assert_equals: 150,150 expected Element node <div class="box"></div> but got Element node <body><div class="box"></div>
<svg height="0">
  <clipPat...
